gcredentials: Consistently use #ifdef for G_CREDENTIALS_* defines
The alternative is to `#define` all the `G_CREDENTIALS_*` defines to
zero in each platform’s definition block, which would result in a
combinatorial explosion of `#define`s (or loads of `#undef`s).

Stick to the convention for GLib and consistently define them to 1 or
leave them undefined.

This helps with `-Wundef` support.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>