docs: Mention how to run the test suite in CONTRIBUTING.md
We don’t want `CONTRIBUTING.md` to become an exhaustive manual of how to
contribute to what is a fairly standard C project using Meson, but it is
a fairly helpful hint to mention how to run the test suite. Then people
can use `meson test --help` and the Meson documentation to learn more.

Pointed out on !4589.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>