gatomic: Mark pointer as (type gpointer)
Otherwise, it appears as an "in" parameter with
  <type name="gint" c:type="volatile gint*"/>
which is very confusing to the bindings.

An alternative would be to mark it as an "inout" integer, which matches
the semantics somewhat. However, that wouldn't quite capture the concept
of the value being modified having (pointer) identity, which is of
utmost importance for atomic operations. Hence, just say that we accept
a pointer to the value.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>