glib: add _g_assert() macro
For example, if a static function has only one caller, then the reviewer
can maybe trivially confirm invariants about the parameters. On the
other hand, if the function is called at many places and form other
files, that may be less trivial.

Assertions are always supposed to hold. But in some cases
they more obviously hold than in others.

But such assertions can still be useful, because they help catching
errors during refactoring and they act as executed comments about the
invariants.

But if they are always present even in release mode, there is an
argument to avoid the overhead of checking the condition.

Also, some assertions conditions may be more expensive to evaluate, and
be avoided for that in release mode.

The usual solution was:

  #ifdef G_ENABLE_DEBUG
     g_assert (condition);
  #endif

but that is cumbersome. It also hides the condition from the compiler in
certain modes, which can lead to unused variables warnings.

The solution now is

    _g_assert (condition);