gparamspecs: Use standard min/max constants rather than literals
This makes the code a little clearer. In most cases, it’s not a
functional change.

In a few cases, the values are different. I believe the original values
were incorrect (accidentally transposed, perhaps). This never caused an
issue because they were all immediately overwritten during construction
of a `GParamSpec`: these values were defaults in the `instance_init`
vfunc of the `GTypeInstance` for a `GParamSpec`, but the actual min/max
for the `GParamSpec` instance were immediately written over them in the
constructor (such as `g_param_spec_int()`).

Spotted in !4593.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>