gbinding: drop obsolete code in weak_unbind()
At the time when this code was added ([1]), the code and the comment was
correct. g_object_run_dispose() did not clear GWeakRef.

That was later adjusted to clear them ([2]), but at various times it was
not ensured that the GWeakRef was cleared *before* the weak notification
is emitted.

This is now fixed, and the checks for "where_the_object_was" are no
longer necessary. Drop them.

I considered to keep the checks just to be extra safe. But we need to
rely on how g_object_run_dispose() works in detail. By now there is a
test that checks GWeakRef are cleared before emitting the notifications,
so we should not accidentally mess this up and the code is no longer
needed.

[1] commit e82eb490fea3 ('Handle the case of g_object_run_dispose() in GBinding')
[2] commit a7262d63576c ('gobject: Cleanup weak locations data as part of dispose')