gbinding: acquire strong reference before taking lock
Acquiring a reference can emit a toggle notification. For the
unref case, we already do that outside the lock (for that reason).
Also acquire the reference before the lock.

It is also slightly better to do fewer things while holding the lock.

Wait. You might think that unbind_internal_locked() resets the weak references
to NULL, while holding the lock. So surely we must not acquire the weak
reference outside the lock, when (a racing) unbind_internal_locked() might
reset it. Indded, with this change, we now might now call
unbind_internal_locked() with a source/target while it's weak reference is
already cleared. However, this is no problem, because we don't do anything to
those objects that wouldn't correctly cope with that. For example, the check

      if (binding->source_notify != 0)
        {
          g_signal_handler_disconnect (source, binding->source_notify);

          g_object_weak_unref_full (source, weak_unbind, context, FALSE);

          binding->source_notify = 0;
        }

ensures that we only g_signal_handler_disconnect() once(*).

In other words, this is fine because all the things we do inside
unbind_internal_locked() do not depend on whether the GWeakRef agrees on
whether we have a source/target object. So we can access the GWeakRef
before taking the lock.

(*) this is not entirely true. A concurrent g_object_run_dispose() call
can hit at any time. It will run g_object_real_dispose(), which calls
g_signal_handlers_destroy(). If that happens at the wrong time, the
g_signal_handler_disconnect() here will misbehave. But that is a
separate problem.