gbinding: don't use GArcBox for BindingContext
GArcBox doesn't seem a great data structure to me. Does it really
simplify the concept of having a ref-count field?

At least for a low-level implementation detail inside GBinding, the
overhead seems unwarrented. GArcBox is 3*sizeof(gsize) of extra memory,
most likely more to get proper alignment. Contrary to a "gatomicrefcount
ref_count" field, which (on amd64) fits in an existing gap in the struct
alignment of BindingContext.

Or point in case, why do we still have

  static BindingContext *
  binding_context_ref (BindingContext *context)
  {
    return g_atomic_rc_box_acquire (context);
  }

The answer is that g_atomic_rc_box_acquire() takes any gpointer, so
binding_context_ref() gives some type safety and readability. But
what is the benefit of GArcBox in this case?