GAysncQueue: use transfer full annotations
Fixes https://gitlab.gnome.org/GNOME/glib/-/issues/3698

This patch marks the data arguments of the push() methods and the return
values of the pop() methods as transfer full.

This patch also marks the return values of the non-timed pop() functions
as non-nullable, as they can only return NULL when assertions fail.

In the process I also found g_async_queue_push_sorted_unlocked() was
missing a non-nullable annotation, so I added it. Given that
g_async_queue_push_sorted() has a non-nullable annotation, it makes no
sense for the unlocked version to lack it.