docs: Stop hiding the Unix-like APIs which are in Gio-2.0.gir
There are four `Unix.+` classes in `Gio-2.0.gir` which need to be
exposed in the `Gio-2.0.gir` docs because they are actually now
cross-platform (which is a move which has caused a lot of pain).

Change the code which filters out the rest of the `Unix.+` classes to
ignore these ones. The rest of the classes continue to be documented via
`GioUnix-2.0.gir`.

Changing the regexs for this involved a fun use of negative lookahead.

See https://gitlab.gnome.org/GNOME/glib/-/issues/3697#note_2459405

Signed-off-by: Philip Withnall <pwithnall@gnome.org>
Helps: #3697