glib: add private _g_bit_lock_init() for fast initialization of locked bitlock
There will be case where we allocate a new structure that has a bitlock
flag. We want to lock on that flag right away (before the pointer is yet
shared to other threads).

We would usually call g_bit_lock(), but as we are inside glib itself, we
*know* that g_bit_lock() will do nothing of relevance except
(atomically) setting the lock bit. Note that we don't need the atomic
setting, as the pointer is not yet shared.

So we can do something cheaper: directly set the bit.

Add _g_bit_lock_init() for that.