gspawn: Improve docstring for g_spawn_async()
Based on my reading of the docs, g_spawn_async_with_pipes_and_fds()
seems to have a much more comprehensive explanation of itself than
g_spawn_async_with_pipes(). As such, it makes sense to me to refer to it
rather than the latter when suggesting where to go for further reading.

Signed-off-by: Stefan Hansson <newbyte@postmarketos.org>