GWin32AppInfo: Add support for startup-notification-id
GdkAppLaunchContext for Wayland / X11 fetch an opaque token
from the compositor.

On Windows, the system doesn't give you any token. Instead,
the token is implicit, and you can pass that to other
processes via dedicated APIs like AllowSetForegroundWindow.

Due to that, we use an internal format for startup-notify-id
that lets us check what to pass to activated apps.