gmessages: Fix win32_keep_fatal_message regression
The functionality of win32_keep_fatal_message was lost with commit
fb9df27776b116d9e8552c0b7b3109245a9c0d26, which introduced the usage of
fputs instead of write. The write system call was wrapped with a macro
for Windows systems to keep fatal messages in an extra buffer.

For the record, the unused dowrite function was removed in commit
5400f4e128b01339b4ae50d8ec52b2182ffa6974.

Reintroduce the functionality to fix this regression in a memory-safe
way.